Controller: check for valid values

+++ controller/inc/class.uicheck_list.inc.php
@@ -1994,7 +1994,7 @@
 			$ok = true;
 
 			$required_actual_hours = isset($config->config_data['required_actual_hours']) && $config->config_data['required_actual_hours'] ? $config->config_data['required_actual_hours'] : false;
-			if ($check_list_status == controller_check_list::STATUS_DONE && $required_actual_hours && (int)$check_list->get_billable_hours() == 0)
+			if ($check_list_status == controller_check_list::STATUS_DONE && $required_actual_hours && (float)$check_list->get_billable_hours() == 0)
 			{
 				phpgwapi_cache::message_set(lang("Please enter billable hours"), 'error');
 				$ok = false;
